+++ packages/component/src/layout/content.tsx
@@ -1,5 +1,15 @@
 import { FC, PropsWithChildren } from "react";
+import { cn } from "..";
+import { LayoutBaseProps } from "./types";
 
-export const Content: FC<PropsWithChildren> = (props) => {
-  return <div>{props.children}</div>;
+export interface ContentProps extends LayoutBaseProps {
+  // fixed?: boolean;
+  // asChild?: boolean;
+}
+
+export const Content: FC<PropsWithChildren<ContentProps>> = (props) => {
+  const { className, ...rest } = props;
+  return <div {...rest} className={cn("orderly-flex-1", className)} />;
 };
+
+Content.displayName = "Content";