+++ packages/eslint-config-custom/CHANGELOG.md
@@ -1,5 +1,38 @@
 # eslint-config-custom
 
+## 0.0.34-alpha.7
+
+### Patch Changes
+
+- fix bug
+
+## 0.0.34-alpha.6
+
+### Patch Changes
+
+- fix bug
+
+## 0.0.34-alpha.5
+
+### Patch Changes
+
+- fix bug
+- fix bug
+- fix bug
+
+## 0.0.34-alpha.4
+
+### Patch Changes
+
+- fix bugs
+- fix bugs
+
+## 0.0.34-alpha.3
+
+### Patch Changes
+
+- fix bug
+
 ## 0.0.34-alpha.2
 
 ### Patch Changes